new mem(heap) controller

+++ Engine/Inc/TimeManager.hpp
@@ -10,12 +10,13 @@ namespace engine {
 
 		void        Update();
 		uint32_t    GetDTms() const;
-		float       GetFPS() const;
+		double      GetFPS() const;
+
 
 	private:
 
-		std::chrono::steady_clock::time_point    m_PreviusT;
-		std::chrono::steady_clock::duration        m_DeltaT;
+		std::chrono::steady_clock::time_point   m_PreviusT;
+		std::chrono::steady_clock::duration     m_DeltaT;
 
 	};
 }
